LED-85 refactor: update user repository interface to return UserSnapshot, streamline imports in user-related files, and enhance test case descriptions for better clarity
